+++ app/controllers/dashboards_controller.rb
@@ -64,10 +64,9 @@ class DashboardsController < ApplicationController
         ping = request.original_url.include?(GibctSiteMapper::PRODUCTION_HOST)
         GibctSiteMapper.new(ping: ping)
 
-        if Settings.institutions.archive
+        if Settings.archiver.archive
           # Archive previous versions of generated data
-          InstitutionsArchive.archive_previous_versions
-          ZipcodeRateArchive.archive_previous_versions
+          Archiver.archive_previous_versions
         end
       else
         flash.alert = 'Production data not updated, remains at previous production version'
