+++ config/settings.yml
@@ -1,5 +1,5 @@
 active_record:
   batch_size: 5000
 
-institutions:
+archiver:
   archive: true
