+++ spec/controllers/dashboards_controller_spec.rb
@@ -128,8 +128,7 @@ RSpec.describe DashboardsController, type: :controller do
   describe 'GET push' do
     before(:each) do
       allow_any_instance_of(GibctSiteMapper).to receive(:ping_search_engines)
-      allow(InstitutionsArchive).to receive(:archive_previous_versions).and_return(nil)
-      allow(ZipcodeRateArchive).to receive(:archive_previous_versions).and_return(nil)
+      allow(Archiver).to receive(:archive_previous_versions).and_return(nil)
     end
     login_user
